Early crack at checkout process.
- Added checkout menu item
- Added stripe_checkout_form
- Added helper functions to pass an amount, in cents, to the checkout
form

Example usage could be:
stripe_set_amount(50.45); // The float value for the amount to charge
sets a SESSION variable in CENTS
drupal_goto('checkout'); // Presents a checkout form to the customer
and confirms the amount. The form injects the necessary JS and the
submit handler processes the card.

Short term TODO:
* Consider security vulnerabilities of setting an amount in the session
variable.
* In the event of creating a customer profile on Stripe's side,
consider a way to tie that ID value to Drupal's UID, so they can check
out in one click.